requirements.txt
aiosqlite==0.21.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
